requirements.txt
# Core dependencies
gradio==3.50.2
python-dotenv==1.0.0
tqdm==4.66.1

# Video and audio processing
ffmpeg-python==0.2.0
moviepy==1.0.3
pydub==0.25.1

# Speech recognition and text-to-speech
assemblyai==0.15.1
gTTS==2.3.2

# Translation and subtitle handling
deep-translator==1.9.2
pysrt==1.1.2

# Utility packages
loguru==0.7.2
